Extract type-family plugin builders
Add AbstractTypePluginBuilder and three concrete subclasses covering
the project-type plugin shapes:

- SingleTypePluginBuilder: one project-type binding, class-style Kotlin
  rendering
- ReifiedSingleTypePluginBuilder: one binding, Kotlin reified-style
  rendering (Java rendering inherited from SingleTypePluginBuilder)
- MultiTypePluginBuilder: multiple project-type bindings (Java only —
  Kotlin rendering not currently exercised by any fixture)

Move the type-family helpers (service injection, build-model
implementation rendering, property return-type, mapping/display
delegators) onto AbstractTypePluginBuilder.
